requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-google-vertexai
google-cloud-aiplatform
langchain-google-genai
google-generativeai
langchain-ollama
-e .
